Factor out ChallengeDetailsReviewPane from GameAcceptModal (for later reuse)
